Disabled lines causing infinite loop bug.
The use of the 'useNavigate' hook is causing an infinite error loop when attempting to redirect after logging in/registering/logging out. For now, these lines have been commented out, and the cause of this bug will be determined later. Automatic redirection has been disabled until the cause of the issue can be determined. For now, manual redirection using the NavBar is required after logging in/registering/logging out.