fixes issues caused by previous merge from Testing to Main
The branch Testing was quite a few commits behind Main. When the merge was done, some issues from the Main branch were preserved, as the fixes done in Testing branch were made in different files, so the code that was causing issues in the previous version of Main were not removed when the pull request was merged, causing some functionality to be broken, even though it was properly functioning in the Testing branch.

This commit ensures that the changes made in the Testing branch are carried over to main.