Recommendation View on user profile
This code update allows users to click on the user profile and view the recommendation. It does so by getting the recommendation information from the recommendation database in the form of an id for the recipient and sender and then using ConnectionCard file it converts those IDs into values to display on the user profile for the recommendations.
This update allows for viewing received and given recommendations.