connect required documents frontend to backend
implementation towards [US-35] #117

This commit adds the required documents functionality to the backend (serializer and JobListing views), as well as to the front end ( Create/Edit Job forms, jobActions and jobsPage).

Required Documents can be configured while creating a Job listing, and edited after the listing has been created. The Job's required documents appear in the job summary on the Jobs screen.