Implementation of external Job Listings
As mentioned in #116, regarding the option for Job Posts to link to third party websites:

This commit implements the feature of having a Job Listing for an external job listing, adding the ability to distinguish between an Internal or External listing, as well as provide a link to the external application in the event that it is an External Job Listing.

This commit includes both the frontend and backend for the above mentioned functionality.