[US-40-41-V2] Job Search: Implement individual job page
This commit provides implementation of an individual job page, which can be used to view a Job from the Job Search [US-40-41-V2] #170 results.

This commit adds a screen for view all Jobs (screens/JobsScreen.js) and a component to render the Jobs (components/Jobs.js), as well as a screen to view an individual Job (screens/JobScreen.js) and a component to render an individual Job (components/Job.js).

This commit also provides the backend implementation for a view which retrieves the details of 1 JobListing, and registers it in backend/urls.py.