[US-32] Connections Newsfeed frontend/backend connection
This commit makes the necessary changes to the Newsfeed view and profile serializer to implement the Connections Newsfeed.

The frontend Newsfeed component was edited to accommodate the changes.

Additionally, the styling of the newsfeed component was updated to better match the changes to the styling changes to the Job component.