add admin authentication to admin views
This commit adds the 'staff_member_required' decorator to the admin views, and passes the access token to the header of the corresponding frontend axios requests. This ensures that only users who are logged in as Admin can execute the admin views.

Additionally, the access token refresh time in settings.py is increased from 5 to 15 minutes.