[maintenance] remove comment elements from Job/Jobs components
This commit removes the Comment boxes and Comment/Like buttons that were present on the Job.js and Jobs.js components, as commenting/liking Jobs is not going to be implemented, and removes unnecessary imports that were present in these files.
Additionally, this commit updates the styling of the Apply button on the individual Job components to match the styling used on the Jobs component.