Fix: [US-48][US-49] Other documents upload fails
This commit fixes the bug mentioned in #188, where if a document is uploaded through the 'Other documents' file input field on the Job Application form. This is fixed by limiting the other documents input field to 1 file. In the backend, the 'other documents' field of the JobApplication model is only set up to accept 1 file. Rather than change the model implementation, we will be limiting the input to 1 other document.