Fix: DM fails to send, uses undefined value
This commit fixes the bug described in issue #189, where on initial load of the Message Inbox page, if you do not click on a conversation in the sidebar and attempt to send a message in the chat window that is automatically open on page load, the message fails to send due to an undefined value being used in the api path.

This is fixed by setting a new state variable, initial, to true when the page loads. If initial is true, meaning the page has just loaded, the activeContactIndex function will return the active index as being 0 (the first conversation in the list) and sets initial to false. On subsequent calls of the activeContactIndex function, it will return the index using the logic that was already present.

fixes #189